Add driver abstraction and oslo_messaging support
Adds a driver abstraction and update modules to use the abstraction
(mainly client.py, server.py, zmq.py).

The abstraction is more RPC based instead of the zmq specific socket
communication. Instead of socket handling in client.py/server.py the
socket handling is now isolated in the zmq driver.

The encode/decode of strings into bytes of the driver input/output is
now also isolated in the driver as not all drivers will require byte
communication.

Also, adds a driver for oslo_messaging, with the aim of being able to
support AMQP with any driver that oslo_messaging that supports.
Initially, only the amqp driver (qpid-proton) is supported.

Signed-off-by: James Slagle <jslagle@redhat.com>