Add customization args to systemd commands
The --service-file-dir arg can be used to override the default directory
(/etc/systemd/system) used to create the service file.

The --config-file-dir arg can be uesd to override the default directory
(/etc/directord) used to create the configuraiton file.

The --binary-path arg is used to override the path to the directord
binary in the generated systemd unit files.

These are needed for packaging so that the files can be generated within
a spec file and placed in the packaged locations.

Signed-off-by: James Slagle <jslagle@redhat.com>